`.json` describing ProHealth Publications
All ProHealth publications may be described in the `_data/publications.json`
file, which will be used to display them on the publications page.

This adds three additional `.json` files which will be populated with content
at runtime. `_collections/_json/connelly.json`, `siek.jsonl`, and
`clawson.json` will be created based on the contents of the main file. The
intention here is that these may be stored for use on people's personal home
pages, or content can be built based on the publications of individual faculty
members without too much difficulty.