Changing `people.json` to `people.yml` and adding Bio's
The previous method for storing information about people (in .json)
proved to be a bit "un-elegant" for keeping track of a person's bio.
This would likely have proved annoying if someone wanted to edit
their bio only to find that it was stored in a single line.

For this reason, `yml` will likely be more maintainable in the future.